A TDD / Web development / Django tutorial
Love the show!

I apologise if naked self-promotion is not the done thing in re: submissions to this podcast, but I humbly submit a link to a book which I am writing about TDD with Django.

It's being published commercially by O'Reilly, but I have convinced them to allow me to make it into a community resource as well, by getting permission to publish a CC-licensed version once the book itself becomes available in print format.  It's already available in "Early Release" form, for free as well as paid.

So far I have 13 chapters available, out of a target 20, and it covers all sorts of TDD, Django and web-dev goodness:
- Selenium
- Django basics: models, views, templates, forms, validation
- CSS + static files
- Deployment! IN chapter 8 I ask readers to *actually* register a domain name, buy a VPS server, install nginx + gunicorn, and push their example app up to the internets.
- AND JAVASCRIPT!   aaaaargh!

Very happy indeed to do an interview on a future podcast, if you think my ill-informed ramblings might be worthy of some sort of public rebuttal by more informed minds such as your own.  I've a knack for saying stupid things which allows the wise to edumecate me.  Here's Gary Bernhardt schooling me on test speed, for example:
http://www.obeythetestinggoat.com/fast-tests-useless-hot-lava-be-damned.html

And here's Russell Keith-Magee fixing my wrong-headed tests for CBGVs:

http://www.obeythetestinggoat.com/fast-tests-useless-hot-lava-be-damned.html

So hopefully we'd be able to have an interesting chat about whether striving for fast tests is a good idea, whether or not the Django Test runner sucks, how to test class-based generic views, how awesome selenium is, whether LiveServerTestCase is hopelessly broken, whether py.test is the future, how you might write a set of tests for a book about testing, and so on....